test: move RealCollector-dependent tests from test/unit to test/integration
context_propagation_test and otlp_grpc_retry_test invoke the external
otelcol binary via RealCollector, so a plain 'dart test test/unit' on a
machine without the collector crashes. Move them to test/integration,
where the other collector-dependent suites live. CI is unaffected:
tool/test.sh runs the whole ./test tree.

Addresses the external-daemon half of #55; the loopback-socket
standardization half remains open.
